% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{add_arm_info}
\alias{add_arm_info}
\title{add_arm_info}
\usage{
add_arm_info(myGR, assembly)
}
\arguments{
\item{myGR}{a GRanges object}

\item{assembly}{an assembly (e.g. hg38)}
}
\value{
A GRanges object with arm information added
}
\description{
Add arm information to a GRanges object
}
\details{
This function adds arm information (e.g. 1q, 17p) to a GRanges object.
}
\examples{
add_arm_info(GenomicRanges::GRanges(seqnames=c("1", "17"),
                                              ranges=IRanges::IRanges(start=c(150e6, 1),
                                                                      end=c(200e6, 20e6))),
                       "hg38")
}
\author{
tlesluyes
}
