% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_week_month.R
\name{add_week_month}
\alias{add_week_month}
\title{根据给定数据框中的日期列,添加月份和周几}
\usage{
add_week_month(df, col_name, format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{df}{数据框}

\item{col_name}{字符串,指定的日期列名}

\item{format}{指定日期列的格式, 同as.Date}

\item{...}{同as.Date}
}
\value{
在原有的数据集上增加两列,weekdays,months 列且为因子
}
\description{
根据给定数据框中的日期列,添加月份和周几
}
\examples{
myfun_example <- function(path = NULL) {
  if (is.null(path)) {
    system.file("extdata", package = "myfun")
  } else {
    system.file("extdata", path, package = "myfun", mustWork = TRUE)
  }
}
path = myfun_example()
df = read_csv_files(path,full.names = TRUE)
head(df)

df_head = head(df,10)
df2 = add_week_month(df_head, '日期')
head(df2)

}
