% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggsec.R
\name{aggsec}
\alias{aggsec}
\title{按行按列合并制定的部门}
\usage{
aggsec(ans, aggcol = c("01T02", "03", "05T06", "07T08", "09"), after = "A-B")
}
\arguments{
\item{aggcol}{指定要合并的部门}

\item{after}{合并后部门的名字}
}
\description{
按行按列合并制定的部门
}
