% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGR.R
\name{cleanGRlistMetadata}
\alias{cleanGRlistMetadata}
\title{cleanGRlistMetadata}
\usage{
cleanGRlistMetadata(myGRList, metadataCols, checkOverlaps = TRUE)
}
\arguments{
\item{myGRList}{a list of harmonized GRanges objects (all granges must be identical across samples)}

\item{metadataCols}{a character vector of metadata column names to check for missing values}

\item{checkOverlaps}{a boolean, whether to check for overlapping ranges in the GRanges objects (default: TRUE)}
}
\value{
A list of cleaned GRanges objects with the same regions but only those with complete metadata across all samples
}
\description{
Remove regions with missing metadata in at least one sample from a list of GRanges objects
}
\details{
This function removes regions that have missing metadata in at least one sample from a list of GRanges objects.
}
\examples{
GR1 <- GenomicRanges::GRanges(seqnames=rep("1", 3),
                              ranges=IRanges::IRanges(start=c(1, 1001, 10001),
                                                      end=c(1000, 10000, 20000)),
                              CNstatus=c(NA, "2+1", "1+1"))
GR2 <- GenomicRanges::GRanges(seqnames=rep("1", 2),
                              ranges=IRanges::IRanges(start=c(500, 10001),
                                                      end=c(10000, 25000)),
                              CNstatus=c("2+1", "1+1"))
myGRList <- harmonizeGRanges(list(GR1=GR1, GR2=GR2), keepHoles=TRUE)
cleanGRlistMetadata(myGRList, metadataCols="CNstatus")
}
\author{
tlesluyes
}
