% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn2en.R
\name{cn2en}
\alias{cn2en}
\title{Translate English from Chinese}
\usage{
cn2en(drv, rawtext, website = "http://fanyi.youdao.com/")
}
\arguments{
\item{drv}{object from selenium module in Python. See examples.}

\item{rawtext}{Chinese which is translated}

\item{website}{don't change it.}
}
\description{
Translate English from Chinese
}
\details{
You should install selenium module in python.
}
\examples{
library(reticulate)
rawtext <- '鎴戞槸涓浗浜戼㹡'
sl <- import('selenium')
drv <- sl$webdriver$Chrome()
cn2en(drv, rawtext)
drv$quit()
}
