% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_BAF_purity_ploidy.R
\name{computeFit}
\alias{computeFit}
\title{computeFit}
\usage{
computeFit(logR, BAF, nMajor, nMinor, gamma, digits = 4)
}
\arguments{
\item{logR}{the logR value of the segment}

\item{BAF}{the BAF value of the segment (upper band only so the value should be in the 0.5-1 space)}

\item{nMajor}{the number of copies of the major allele}

\item{nMinor}{the number of copies of the minor allele}

\item{gamma}{the gamma parameter is platform-dependent and represents the expected logR decrease in a diploid sample where one copy is lost (should be 1 for HTS data and 0.55 for SNP arrays)}

\item{digits}{a numeric, the number of digits to round to (default: 4)}
}
\value{
A list with the rho (=purity), psi (=total ploidy) and psit (=tumour ploidy) values
}
\description{
Compute the purity/ploidy fit for a given segment
}
\details{
This function computes the purity/ploidy fit (rho, psi and psit) for a given segment (from logR, BAF, proposed nMajor and proposed nMinor).
}
\examples{
# A segment has logR=0.5361 and BAF=0.3448/0.6552
# What is the purity/ploidy fit if I believe that the segment is 2+1?
computeFit(0.5361, 0.6552, 2, 1, 1) # purity=90\%; ploidy=2
}
\seealso{
https://doi.org/10.1038/s41592-020-01013-2
}
\author{
tlesluyes
}
