% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeISA.R
\name{computeISA}
\alias{computeISA}
\title{computeISA}
\usage{
computeISA(GR1, GR2, CNstatus = "CNstatus")
}
\arguments{
\item{GR1}{a GRanges object corresponding to a single CNA profile}

\item{GR2}{a GRanges object corresponding to a single CNA profile}

\item{CNstatus}{a metadata column name for the copy-number status (default: "CNstatus"). Can be total (e.g. "3") or allele-specific (e.g. "2+1")}
}
\value{
A percentage representing the ISA
}
\description{
Compute the inter-sample agreement (ISA)
}
\details{
This function computes the inter-sample agreement (ISA) between two profiles (as GRanges objects). This corresponds to the fraction of the genome (\%) with the same CN status.
}
\examples{
GR1 <- GenomicRanges::GRanges(seqnames=rep("1", 3),
                              ranges=IRanges::IRanges(start=c(1, 1001, 10001),
                                                      end=c(1000, 10000, 20000)),
                              CNstatus=c("1+1", "2+1", "1+1"))
GR2 <- GenomicRanges::GRanges(seqnames=rep("1", 2),
                              ranges=IRanges::IRanges(start=c(500, 10001),
                                                      end=c(10000, 25000)),
                              CNstatus=c("2+1", "1+1"))
# in this example:
#    Region 500-1000 (size=501) is 1+1 for GR1 and 2+1 for GR2
#    Region 1001-20000 (size=19000) is identical between GR1 and GR2 (both 2+1 and 1+1)
#    ISA is: 19000/19501 = 97.43\%
computeISA(GR1, GR2)
}
\author{
tlesluyes
}
