% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeISA.R
\name{computeISA_batch}
\alias{computeISA_batch}
\title{computeISA_batch}
\usage{
computeISA_batch(myGRList, cores = 1, min_seg_size = 0, CNstatus = "CNstatus")
}
\arguments{
\item{myGRList}{a list of GRanges objects, each object should correspond to one CNA profile}

\item{cores}{a numeric, the number of cores to use (default: 1)}

\item{min_seg_size}{a numeric, the minimum segment size (in bp) to consider (default: 0)}

\item{CNstatus}{a metadata column name for the copy-number status (default: "CNstatus"). Can be total (e.g. "3") or allele-specific (e.g. "2+1")}
}
\value{
A matrix of ISA values
}
\description{
Compute the inter-sample agreement (ISA) for a batch of samples
}
\details{
This function computes the inter-sample agreement (ISA) between multiple profiles (as a list of GRanges objects).
}
\examples{
require("GenomicRanges")
GR1=GRanges(seqnames=rep("1", 3),
            ranges=IRanges(start=c(1, 1001, 10001), end=c(1000, 10000, 20000)),
            CNstatus=c("1+1", "2+1", "1+1"))
GR2=GRanges(seqnames=rep("1", 2),
            ranges=IRanges(start=c(500, 10001), end=c(10000, 25000)),
            CNstatus=c("2+1", "1+1"))
GR3=GRanges(seqnames="1",
            ranges=IRanges(start=500, end=25000),
            CNstatus="1+1")
myGRList=list(GR1, GR2, GR3)
names(myGRList)=c("GR1", "GR2", "GR3")
computeISA_batch(myGRList)
}
\author{
tlesluyes
}
