% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_BAF_purity_ploidy.R
\name{computeLogR}
\alias{computeLogR}
\title{computeLogR}
\usage{
computeLogR(nMajor, nMinor, purity, ploidy, digits = 4)
}
\arguments{
\item{nMajor}{the number of copies of the major allele}

\item{nMinor}{the number of copies of the minor allele}

\item{purity}{the purity estimate of the tumour}

\item{ploidy}{the ploidy estimate of the tumour}

\item{digits}{a numeric, the number of digits to round to (default: 4)}
}
\value{
A number representing the logR value
}
\description{
Compute the theoretical logR value for a given segment
}
\details{
This function computes the theoretical logR value for a given segment (from nMajor, nMinor, purity and ploidy values). Since logR isn't allele-specific, ntot can be used instead of nMajor (and nMinor should set to 0).
}
\examples{
# A 2+1 state in a diploid tumour with 90\% purity
computeLogR(2, 1, 0.9, 2)
# A loss of 1 copy (2+1) in a pseudo-tetraploid tumour with 60\% purity
computeLogR(2, 1, 0.6, 3.5)
}
\seealso{
https://doi.org/10.1038/s41592-020-01013-2
}
\author{
tlesluyes
}
