% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMD.R
\name{computeMD}
\alias{computeMD}
\title{computeMD}
\usage{
computeMD(GR1, GR2, nMajor = "nMajor", nMinor = "nMinor", convertMb = FALSE)
}
\arguments{
\item{GR1}{a GRanges object corresponding to a single CNA profile}

\item{GR2}{a GRanges object corresponding to a single CNA profile}

\item{nMajor}{a metadata column name for the major allele (default: "nMajor")}

\item{nMinor}{a metadata column name for the minor allele (default: "nMinor")}

\item{convertMb}{a boolean, the MD will be converted to megabases if set to TRUE (default: FALSE)}
}
\value{
A numeric value representing the MD
}
\description{
Compute the Manhattan distance (MD)
}
\details{
This function computes the Manhattan distance (MD) between two profiles (as GRanges objects).
}
\examples{
GR1=GenomicRanges::GRanges(seqnames=rep("1", 3),
                           ranges=IRanges::IRanges(start=c(1, 1001, 10001),
                                                   end=c(1000, 10000, 20000)),
                           nMajor=c(1, 2, 1),
                           nMinor=c(1, 1, 1))
GR2=GenomicRanges::GRanges(seqnames=rep("1", 2),
                           ranges=IRanges::IRanges(start=c(500, 10001),
                                                   end=c(10000, 25000)),
                           nMajor=c(2, 1),
                           nMinor=c(1, 1))
# in this example:
#    Region 500-1000 (size=501) is 1+1 for GR1 and 2+1 for GR2
#    Region 1001-20000 (size=19000) is identical between GR1 and GR2 (both 2+1 and 1+1)
#    MD is: (abs(2-1)+abs(1-1))*501 = 501
computeMD(GR1, GR2)
}
\author{
tlesluyes
}
