% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMD.R
\name{computeMD_batch}
\alias{computeMD_batch}
\title{computeMD_batch}
\usage{
computeMD_batch(
  myGRList,
  cores = 1,
  min_seg_size = 0,
  nMajor = "nMajor",
  nMinor = "nMinor",
  convertMb = FALSE
)
}
\arguments{
\item{myGRList}{a list of GRanges objects, each object should correspond to one CNA profile}

\item{cores}{a numeric, the number of cores to use (default: 1)}

\item{min_seg_size}{a numeric, the minimum segment size (in bp) to consider (default: 0)}

\item{nMajor}{a metadata column name for the major allele (default: "nMajor")}

\item{nMinor}{a metadata column name for the minor allele (default: "nMinor")}

\item{convertMb}{a boolean, the MD will be converted to megabases if set to TRUE (default: FALSE)}
}
\value{
A matrix of MD values
}
\description{
Compute the Manhattan distance (MD) for a batch of samples
}
\details{
This function computes the Manhattan distance (MD) between multiple profiles (as a list of GRanges objects).
}
\examples{
require("GenomicRanges")
GR1=GRanges(seqnames=rep("1", 3),
            ranges=IRanges(start=c(1, 1001, 10001), end=c(1000, 10000, 20000)),
            nMajor=c(1, 2, 1),
            nMinor=c(1, 1, 1))
GR2=GRanges(seqnames=rep("1", 2),
            ranges=IRanges(start=c(500, 10001), end=c(10000, 25000)),
            nMajor=c(2, 1),
            nMinor=c(1, 1))
GR3=GRanges(seqnames="1",
            ranges=IRanges(start=500, end=25000),
            nMajor=1,
            nMinor=1)
myGRList=list(GR1, GR2, GR3)
names(myGRList)=c("GR1", "GR2", "GR3")
computeMD_batch(myGRList)
}
\author{
tlesluyes
}
