% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_CNAs.R
\name{example_CNAs}
\alias{example_CNAs}
\title{example_CNAs}
\usage{
example_CNAs(n = 1, assembly = "hg38", seed = 1234)
}
\arguments{
\item{n}{an integer, the number of CNA profiles to generate (default: 1)}

\item{assembly}{a character, the assembly to use (default: "hg38")}

\item{seed}{an integer, the seed for reproducibility (default: 1234)}
}
\value{
A list of GRanges objects, each corresponding to one CNA profile
}
\description{
Generate random CNA profiles
}
\details{
This function generates random CNA profiles for a given assembly.
}
\examples{
example_CNAs(n=3, assembly="hg38")
}
\author{
tlesluyes
}
