% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeGRanges.R
\name{excludeGRanges}
\alias{excludeGRanges}
\title{excludeGRanges}
\usage{
excludeGRanges(GR.ref, GR.toremove)
}
\arguments{
\item{GR.ref}{a GRanges object to be filtered}

\item{GR.toremove}{a GRanges object containing regions to exclude}
}
\value{
A filtered GRanges object
}
\description{
Exclude GRanges regions
}
\details{
This function excludes GRanges regions from a reference GRanges object.
}
\examples{
GR1 <- GenomicRanges::GRanges(seqnames = rep(1, 2),
                              ranges=IRanges::IRanges(start=c(1, 5e5+1), end=c(5e5, 1e6)),
                              score=c(3, 2))
GR2 <- GenomicRanges::GRanges(seqnames = 1,
                              ranges=IRanges::IRanges(start=4e5+1, end=6e5))
excludeGRanges(GR1, GR2)
}
\author{
tlesluyes
}
