% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bmatrix.R
\name{fun_bmatrix}
\alias{fun_bmatrix}
\title{把R 中的矩阵或数据框对象转变为 latex 的矩阵格式}
\usage{
fun_bmatrix(x, digits = NULL, ...)
}
\arguments{
\item{x}{矩阵或数据框对象}

\item{digits, }{保留的小数}

\item{...}{其余的参数}
}
\value{
输出 latex 矩阵环境
}
\description{
把R 中的矩阵或数据框对象转变为 latex 的矩阵格式
}
\examples{
fun_bmatrix(iris[1:5,])

}
