% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_describe.R
\name{fun_describe}
\alias{fun_describe}
\title{描述统计}
\usage{
fun_describe(
  x,
  na.rm = TRUE,
  interp = FALSE,
  skew = TRUE,
  ranges = TRUE,
  trim = 0.1,
  type = 3,
  check = TRUE,
  fast = NULL,
  quant = NULL,
  IQR = FALSE,
  omit = FALSE,
  data = NULL
)
}
\arguments{
\item{x}{ A data frame or matrix}

\item{na.rm}{The default is to delete missing data.  na.rm=FALSE will delete the case. }

\item{interp}{Should the median be standard or interpolated}

\item{skew}{ Should the skew and kurtosis be calculated? }

\item{ranges}{ Should the range be calculated? }

\item{trim}{trim=.1 -- trim means by dropping the top and bottom trim fraction}

\item{type}{Which estimate of skew and kurtosis should be used?  (See details.) }

\item{check}{Should we check for non-numeric variables?  Slower but helpful.}

\item{fast}{if TRUE, will do n, means, sds, min, max, ranges for an improvement in speed.  If 
  NULL, will switch to fast mode for large (ncol * nrow > 10^7) problems, otherwise 
  defaults to fast = FALSE}

\item{quant}{if not NULL, will find the specified quantiles (e.g. quant=c(.25,.75) will 
  find the 25th and 75th percentiles)}

\item{IQR}{If TRUE, show the interquartile range}

\item{omit}{Do not convert non-numerical variables to numeric, omit them instead}

\item{data}{Allows formula input for specific grouping variables}
}
\description{
注意 参数 type 涉及几种不同的偏度与峰度,这里全都是 0 为标准正态
}
\seealso{
\link[psych:describe]{psych::describe}
}
