% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_describeBy.R
\name{fun_describeBy}
\alias{fun_describeBy}
\title{分组描述统计}
\usage{
fun_describeBy(
  x,
  group = NULL,
  mat = FALSE,
  type = 3,
  digits = 15,
  data = NULL,
  ...
)
}
\arguments{
\item{x}{a data.frame or matrix. See note for statsBy.  }

\item{group}{a grouping variable or a list of grouping variables.  (may be ignored if calling using the formula mode.)}

\item{mat}{provide a matrix output rather than a list}

\item{type}{Which type of skew and kurtosis should be found}

\item{digits}{When giving matrix output, how many digits should be reported?}

\item{data}{Needed if using formula input}

\item{...}{parameters to be passed to describe}
}
\description{
注意 参数 type 涉及几种不同的偏度与峰度,这里全都是 0 为标准正态
}
\seealso{
\link[psych:describe.by]{psych::describeBy}
}
