% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lag.R
\name{fun_lag}
\alias{fun_lag}
\title{计算滞后阶数}
\usage{
fun_lag(x, n = 1L, default = NULL, order_by = NULL, ...)
}
\arguments{
\item{x}{A vector}

\item{n}{Positive integer of length 1, giving the number of positions to
lag or lead by}

\item{default}{The value used to pad \code{x} back to its original size after the
lag or lead has been applied. The default, \code{NULL}, pads with a missing
value. If supplied, this must be a vector with size 1, which will be cast
to the type of \code{x}.}

\item{order_by}{An optional secondary vector that defines the ordering to use
when applying the lag or lead to \code{x}. If supplied, this must be the same
size as \code{x}.}

\item{...}{Not used.}
}
\description{
计算滞后阶数,要小心,有很多函数需要转换,因为会删除 NA 值,这个函数不会
不太清楚如何直接导入别人包中的函数,怕同名,弄个fun_前缀开始的函数,统一书写
}
\seealso{
\link[dplyr:lead-lag]{dplyr::lag}
}
