% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load.R
\name{fun_load}
\alias{fun_load}
\title{Load Package}
\usage{
fun_load(
  ...,
  char,
  install = TRUE,
  update = getOption("pac_update"),
  character.only = FALSE
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[pacman:p_load]{pacman::p_load}}
  \describe{
    \item{\code{}}{}
  }}

\item{char}{Character vector containing packages to load.  If you are calling
\code{p_load} from within a function (or just having difficulties calling it 
using a character vector input) then pass your character vector of packages 
to load to this parameter directly.}

\item{install}{logical.  If \code{TRUE} will attempt to install a package not 
found in the library.}

\item{update}{logical.  If \code{TRUE} will attempt to update all out of date 
packages.  Default allows the user to set a \code{"pac_update"} in his/her 
.Rprofile.}

\item{character.only}{logical. If \code{TRUE} then \code{p_load} will only 
accept a single input which is a character vector containing the names of 
packages to load.}
}
\description{
加载包,如果该包没有安装,则安装以后再加载
}
\examples{
\dontrun{
p_load(lattice)
p_unload(lattice)
p_load(lattice, foreign, boot, rpart)
p_loaded()
p_unload(lattice, foreign, boot, rpart)
p_loaded()
}
}
\seealso{
\link[pacman:p_load]{pacman::p_load}
}
