% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_normality_tests.R,
%   R/fun_norm_univar_tests.R
\name{fun_normality_tests}
\alias{fun_normality_tests}
\alias{fun_norm_univar_tests}
\title{正太分布检验}
\usage{
fun_normality_tests(df)

fun_norm_univar_tests(x, cn = "")
}
\arguments{
\item{df}{一个数据框, 会对数据框做数值型检查,只检查带数值型的列}

\item{x}{一个数值型向量}

\item{cn}{一个字符串, 与 x 对应的字符串, 只用来修改 qq 图上的 title, 使得我们容易区分,这是那个变量的 qq 图而已}
}
\value{
\code{fun_normality_tests()} 返回一个 list, 每一列的正态分布检验结果存储在list 的元素中, 每个元素是一个矩阵

\code{fun_norm_univar_tests()} 返回一个矩阵, 行是对应正态分布检验的方法,列分别代表 P 值, 统计量,
正态分布的判断结果, 结果为 1 表示接受原假设, 是正态分布.0 则相反
}
\description{
正太分布检验

进行单变量的正态分布检验
}
\section{Functions}{
\itemize{
\item \code{fun_normality_tests()}: 对一个数据框中的数值型做正正态分布检验,主要调用的是 \code{fun_norm_univar_tests()}

\item \code{fun_norm_univar_tests()}: 进行单变量的正态分布检验, 提供几种不同类型的正态检验方法
\itemize{
\item "Shapiro", see \link[stats:shapiro.test]{stats::shapiro.test}
\item "KS" , see \link[stats:ks.test]{stats::ks.test}
\item "AD" , see \link[nortest:ad.test]{nortest::ad.test}
\item "Lillie", see \link[nortest:lillie.test]{nortest::lillie.test}
\item "SF", see \link[nortest:sf.test]{nortest::sf.test}
\item "CVM", see \link[nortest:cvm.test]{nortest::cvm.test}
\item "JB", see \link[moments:jarque.test]{moments::jarque.test}
}

}}
\examples{
data(iris)
fun_normality_tests(iris)

data(iris)
fun_norm_univar_tests(iris[,1],colnames(iris)[1])

}
