% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_repmat_vec.R
\name{fun_repmat_vec}
\alias{fun_repmat_vec}
\alias{fun_repmat}
\title{把向量进行整体填充为矩阵.}
\usage{
fun_repmat_vec(x, nrow, ncol, byrow = FALSE)

fun_repmat(x, nrow, ncol)
}
\arguments{
\item{x}{必须是一个向量, 在 \code{fun_repmat_vec} 中必须是向量, 在\code{fun_repmat} 中必须是矩阵}

\item{nrow}{数字, 行数}

\item{ncol}{数字, 列数}

\item{byrow}{默认是FALSE , 即把传入的 向量 x 看做列向量为整体进行填充, 如果是 TRUE,表示把向量 x 看做行向量为整体进行填充}
}
\value{
返回一个矩阵
}
\description{
由于经常使用向量与矩阵混合计算, 会出现填充问题, 为了更加准确,
建议先把向量转为矩阵以后,再去计算,这样可靠一些. 这个函数参考 matlab 中的 repmat 函数
对向量进行整体的填充
}
\section{Functions}{
\itemize{
\item \code{fun_repmat_vec()}: 把向量作为整体进行填充,传递的 x 必须是向量

\item \code{fun_repmat()}: 把矩阵作为整体进行填充, 传递的 x 必须是矩阵

}}
\examples{
fun_repmat_vec(c(1:3),2,4)
fun_repmat_vec(c(1:3),4,2,byrow = TRUE)

fun_repmat(matrix(c(1:6),2,3), 2,4)
}
