% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ROC.R
\name{fun_ROC}
\alias{fun_ROC}
\title{计算对数/简单收益率}
\usage{
fun_ROC(x, n = 1, type = c("continuous", "discrete"), na.pad = TRUE)
}
\arguments{
\item{x}{Price, volume, etc. series that is coercible to xts or matrix.}

\item{n}{Number of periods to use.}

\item{type}{Compounding type; either \code{"continuous"} (the default) or
\code{"discrete"}.}

\item{na.pad}{Should periods prior to \code{n} be appended?  Default is
\code{TRUE}.}
}
\description{
试了好多对数收益率的函数,发现有各种要求,这个函数没有,怕记不住
}
\seealso{
\link[TTR:changes]{TTR::ROC}
}
