% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{get_Seqinfo}
\alias{get_Seqinfo}
\title{get_Seqinfo}
\usage{
get_Seqinfo(assembly, genome = NA)
}
\arguments{
\item{assembly}{an assembly (e.g. hg38) or a data.frame with expected cytoband information (chr, size)}

\item{genome}{genome name (e.g. hg38), required if \code{assembly} is a data.frame (default: NA)}
}
\value{
A Seqinfo object
}
\description{
Get Seqinfo object
}
\details{
This function gets Seqinfo object for a given assembly.
}
\examples{
get_Seqinfo("hg38")
}
\author{
tlesluyes
}
