% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFEVD.R
\name{GFEVD}
\alias{GFEVD}
\title{Compute GFEVD of linear VAR by Pesaran and Shin (1998)}
\usage{
GFEVD(ma, sig_u, imp_var, res_var, cholesky = FALSE)
}
\arguments{
\item{ma}{a list, it's MA coeffients from \code{ar2ma}}

\item{sig_u}{a cov matrix from VAR models. Note the order of variables in \code{sig_u}
is same with one of \code{ma[[i]]}.}

\item{imp_var}{a numerical scalor which specifies the impulse variable.}

\item{res_var}{a numerical scalor which specifies the imponse variable.}

\item{cholesky}{logical value. Decomposition of prediction error variance
obtained using Cholesky decomposition or not?}
}
\value{
a vector whose length are horizons.
}
\description{
Compute GFEVD of linear VAR by Pesaran and Shin (1998)
}
\examples{
# see ar2ma function
}
