% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI.R
\name{GI}
\alias{GI}
\title{Generalized IRF}
\usage{
GI(ma, sig_u, imp_var = 1)
}
\arguments{
\item{ma}{a list, it's MA coeffients from \code{ar2ma}}

\item{sig_u}{a cov matrix from VAR models. Note the order of variables in \code{sig_u}
is same with \code{ma}.}

\item{imp_var}{a numerical scalor which specifies the impulse variable.}
}
\value{
a data frame, its row is variables and its coloumn is horizons.
}
\description{
Compute GIRF of linear VAR by Koop et al. (1996)
}
\examples{
# see ar2ma function
}
