% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_cnt.R
\docType{data}
\name{iso_cnt}
\alias{iso_cnt}
\title{ISO code for countries}
\format{
A data frame with 247 rows and 9 variables
\describe{
  \item{iso2c}{2 bit code for countries}
  \item{iso3c}{3 bit code for countries}
  \item{number}{}
  \item{country_en}{English names for countries}
  \item{country_cn}{Chinese names for countries}
}
}
\usage{
iso_cnt
}
\description{
A data containing the countrie' English and Chinese names
}
\keyword{datasets}
