% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{load_CHRsize}
\alias{load_CHRsize}
\title{load_CHRsize}
\usage{
load_CHRsize(assembly)
}
\arguments{
\item{assembly}{an assembly (hg19, hg38 or CHM13)}
}
\value{
A data.frame with the \code{CHRsize} information
}
\description{
Load \code{CHRsize} information
}
\details{
This function loads \code{CHRsize} information for a given assembly. It is then available as a data.frame called \code{CHRsize} in the environment.
}
\examples{
load_CHRsize("hg38"); head(CHRsize)
}
\author{
tlesluyes
}
