\name{myFun-package}
\alias{myFun-package}
\alias{myFun}
\docType{package}
\title{
\packageTitle{myFun}
}
\description{
\packageDescription{myFun}
}

\details{
\packageDESCRIPTION{myFun}
\packageIndices{myFun}
}

\author{
\packageAuthor{myFun}

Maintainer: \packageMaintainer{myFun}
}
\references{
  Hadley, Wickham. \emph{Advanced R}. Chapman & Hall/CRC The R Series, 1 edition. September 25, 2014.
}

\keyword{ Utilities }

