% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{normality_tests}
\alias{normality_tests}
\title{正太分布检验}
\usage{
normality_tests(df)
}
\arguments{
\item{df}{一个数据框, 会对数据框做数值型检查,只检查带数值型的列}
}
\value{
list, 返回一个 list, 每一列的正态分布检验结果存储在list 的元素中, 每个元素是一个矩阵
}
\description{
正太分布检验
}
\examples{
data(iris)
normality_tests(iris)

}
