% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_files.R
\name{read_csv_files}
\alias{read_csv_files}
\title{读取某个路径下的所有 csv 文件,并添加一列 symbol, 该列的内容是去除后缀的文件名}
\usage{
read_csv_files(path, pattern = "^\\\\d{6}\\\\.csv$", ...)
}
\arguments{
\item{path}{路径}

\item{pattern}{正则表达式}

\item{...}{同 list.files}
}
\value{
返回一个 data.table
}
\description{
读取某个路径下的所有 csv 文件,并添加一列 symbol, 该列的内容是去除后缀的文件名
}
\examples{
myfun_example <- function(path = NULL) {
  if (is.null(path)) {
    system.file("extdata", package = "myfun")
  } else {
    system.file("extdata", path, package = "myfun", mustWork = TRUE)
  }
}
path = myfun_example()
df = read_csv_files(path,full.names = TRUE)
head(df)
}
