% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_files.R
\name{read_files}
\alias{read_files}
\alias{read_xlsx_files}
\alias{read_csv_files}
\title{读取所有符合要求的csv/xlsx 文件}
\usage{
read_files(file_list, ...)

read_xlsx_files(file_list, ...)

read_csv_files(file_list, ...)
}
\arguments{
\item{file_list}{一个 csv or xlsx 的文件向量, 一般由 \code{list.files()} 产生}

\item{...}{传递给读取 csv/xlsx 的函数的值}
}
\value{
返回一个 data.table
}
\description{
读取所有符合要求的csv/xlsx 文件, 并添加一列列名为 \verb{_file_name}, 该列名是读取文件的文件名(包含后缀)
}
\section{Functions}{
\itemize{
\item \code{read_files()}: 读取给定的所有 csv and xlsx 文件

\item \code{read_xlsx_files()}: 读取给定的所有 xlsx 文件,采用的是\code{readxl::read_xlsx(path = filename)}来读取

\item \code{read_csv_files()}: 读取给定的所有 csv 文件, 采用的是\code{data.table::fread(file = filename)} 来读取

}}
\examples{
myfun_example <- function(path = NULL) {
  if (is.null(path)) {
    system.file("extdata", package = "myfun")
  } else {
    system.file("extdata", path, package = "myfun", mustWork = TRUE)
  }
}
path = myfun_example()
file_list = list.files(path, pattern = '\\\\.csv$',full.names = TRUE)
df = read_csv_files(file_list)
head(df)

file_list = list.files(path, pattern = '\\\\.xlsx$',full.names = TRUE)
df = read_xlsx_files(file_list)
head(df)

file_list = list.files(path, pattern = '(\\\\.xlsx$)|(\\\\.csv)',full.names = TRUE)
df = read_files(file_list)
head(df)
}
\seealso{
\code{\link[base:list.files]{base::list.files()}} \code{\link[=read_csv_files]{read_csv_files()}} \link{read_xlsx_files}
}
