% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_BAF_purity_ploidy.R
\name{reestimate_ploidy}
\alias{reestimate_ploidy}
\title{reestimate_ploidy}
\usage{
reestimate_ploidy(rho.old, psit.old, rho.new, WGD, digits = 4)
}
\arguments{
\item{rho.old}{old purity estimate}

\item{psit.old}{old ploidy estimate}

\item{rho.new}{new purity estimate}

\item{WGD}{number of WGD events (0 if there is no WGD)}

\item{digits}{a numeric, the number of digits to round to (default: 4)}
}
\value{
A number representing the re-estimated ploidy
}
\description{
Compute the re-estimated ploidy for a given sample
}
\details{
This function computes the re-estimated ploidy for a given sample (from its old purity/ploidy fit and the re-estimated purity).
}
\examples{
# A pseudo-diploid sample has purity=74\% and ploidy=2.4
# What is the re-estimated ploidy if I believe that the sample has purity=61\%?
reestimate_ploidy(0.74, 2.4, 0.61, 0)
}
\author{
tlesluyes
}
