% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_BAF_purity_ploidy.R
\name{reestimate_purity}
\alias{reestimate_purity}
\title{reestimate_purity}
\usage{
reestimate_purity(rho.old, psit.old, switch, digits = 4)
}
\arguments{
\item{rho.old}{old purity estimate}

\item{psit.old}{old ploidy estimate}

\item{switch}{a character ("double" or "halve") indicating whether the ploidy should be doubled or halved}

\item{digits}{a numeric, the number of digits to round to (default: 4)}
}
\value{
A number representing the re-estimated purity
}
\description{
Compute the re-estimated purity for a given sample
}
\details{
This function computes the re-estimated purity for a given sample in the context of a jump in ploidy (so the matched ploidy needs to be doubled or halved).
}
\examples{
# A sample has purity=74\% and ploidy=2.4 but the CNA profile needs to be doubled
# What is the re-estimated purity?
reestimate_purity(0.74, 2.4, "double")
}
\author{
tlesluyes
}
