% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackages.R
\name{Rpackages}
\alias{Rpackages}
\title{Rpackages}
\usage{
Rpackages(
  CRAN_URL = "http://cran.us.r-project.org",
  Bioconductor_URL = "https://www.bioconductor.org/packages/release/bioc/"
)
}
\arguments{
\item{CRAN_URL}{the CRAN URL (default: "http://cran.us.r-project.org")}

\item{Bioconductor_URL}{the Bioconductor URL (default: "https://www.bioconductor.org/packages/release/bioc/")}
}
\value{
A data.frame with the installed packages and an additional column: Source (possible values: Base, CRAN, Bioconductor, External)
}
\description{
List installed packages and determine their source
}
\details{
This function lists installed packages and determine whether they are base packages or come from CRAN/Bioconductor or if they are external (GitHub, SourceForge, etc.). This function requires an internet connection.
}
\examples{
head(Rpackages())
}
\author{
tlesluyes
}
