% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{seqinfo2GR}
\alias{seqinfo2GR}
\title{seqinfo2GR}
\usage{
seqinfo2GR(myseqinfo)
}
\arguments{
\item{myseqinfo}{a Seqinfo object}
}
\value{
A GRanges object
}
\description{
Convert Seqinfo to GRanges
}
\details{
This function generates a GRanges object from a Seqinfo object.
}
\examples{
seqinfo2GR(get_Seqinfo("hg38"))
}
\author{
tlesluyes
}
