% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wd.R
\name{set_wd}
\alias{set_wd}
\title{Set the working directory.}
\usage{
set_wd(path = NULL, ask = FALSE)
}
\arguments{
\item{path}{NULL (default) or a specific path. Default is to extract the path of the currently opened file (usually .R or .Rmd) using the rstudioapi::getSourceEditorContext function.}

\item{ask}{TRUE or FALSE (default). If TRUE, you can select a folder with the prompt of a dialog.}
}
\value{
Invisibly return the path.
}
\description{
Set the working directory. If not, create the working directory. You cannot create the directory recursively
}
\examples{
## Not run:
d = getwd()
m = tempdir()
set_wd(m)
set_wd(d)
}
