% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_segmetation.R
\name{summarise_segmetation}
\alias{summarise_segmetation}
\title{summarise_segmetation}
\usage{
summarise_segmetation(DF, col_chr, col_start, col_end, col_values)
}
\arguments{
\item{DF}{a data.frame with segmentation data}

\item{col_chr}{a string, the name of the column containing the chromosome}

\item{col_start}{a string, the name of the column containing the start position}

\item{col_end}{a string, the name of the column containing the end position (can be the same as col_start for SNP-based segmentation where start=end)}

\item{col_values}{a vector of strings, the names of the columns containing the values of interest (logR, BAF, etc.)}
}
\value{
A named list with segments being a data.frame with the summarised information and IDs being a list of SNPs/loci associated with the different segments
}
\description{
Summarise segmentation data
}
\details{
This function summarises segmentation data, typically logR and/or BAF values for individual SNPs or loci.
}
\examples{
DF=data.frame(chr=c(rep("chr1", 10),rep("chr2", 6)),
              pos=c(1:10*1e3, 1:6*1e3),
              logR=c(rep(0, 4), rep(0.54, 3), rep(0, 3), rep(-0.86, 3), rep(0, 3)),
              BAF=c(rep(0.5, 4), rep(0.34, 3), rep(0.5, 3), rep(0.09, 3), rep(0.5, 3)),
              row.names=paste0("SNP_", 1:16))
DF
summarise_segmetation(DF, "chr", "pos", "pos", c("logR", "BAF"))
}
\author{
tlesluyes
}
