% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tran.R
\name{text_tran}
\alias{text_tran}
\title{Texts Translatation}
\usage{
text_tran(rawtext, appid, keyimp, from = "en", to = "zh")
}
\arguments{
\item{rawtext}{a text which is translated}

\item{appid}{from BaiDu}

\item{keyimp}{form BaiDu}

\item{from}{what language \code{rawtext} belong to, and default is \code{'en'} which is English}

\item{to}{the translated target language, and default is \code{'zh'} which is Chinese}
}
\description{
Texts Translatation
}
\details{
call BaiDu App to translate texts. See https://api.fanyi.baidu.com/product/113. Free
version is limited to calling one time by one second.
}
\examples{

# write your appid and key
# text_tran('apple',appid = '', keyimp = '')
}
