% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_norm_tests.R
\name{univar_norm_tests}
\alias{univar_norm_tests}
\title{只能进行单变量的正态分布假设}
\usage{
univar_norm_tests(x, cn = "")
}
\arguments{
\item{x}{一个数值型向量}

\item{cn}{一个字符串, 与 x 对应的字符串}
}
\value{
matrix, 一个矩阵, 行是对应正态分布检验的方法,列分别代表 P 值, 统计量,
正态分布的判断结果, 结果为 1 表示接受原假设, 是正态分布.0 则相反
}
\description{
说明一下,最近发现居然可以写中文了,于是为了方便快捷,反正都是自用,就写中文啦(虽然有警告)
}
\examples{
data(iris)
univar_norm_tests(iris[,1],colnames(iris)[1])

}
