% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGR.R
\name{checkGR}
\alias{checkGR}
\title{checkGR}
\usage{
checkGR(myGR, checkOverlaps = TRUE)
}
\arguments{
\item{myGR}{a GRanges object}

\item{checkOverlaps}{whether to check for overlapping ranges in the GRanges object (default: TRUE)}
}
\value{
TRUE if the input is a valid GRanges object
}
\description{
Check that the given object is a valid GRanges object
}
\details{
This function checks that the given object is a valid GRanges object.
}
\examples{
GR1 <- GenomicRanges::GRanges(seqnames="1",
                              ranges=IRanges::IRanges(start=c(1, 1001), end=c(1000, 2000)))
checkGR(GR1)
GR2 <- GenomicRanges::GRanges(seqnames="1",
                              ranges=IRanges::IRanges(start=c(1, 500), end=c(1000, 1000)))
checkGR(GR2, checkOverlaps=FALSE)
}
\author{
tlesluyes
}
