% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGRlist.R
\name{checkGRlist}
\alias{checkGRlist}
\title{checkGRlist}
\usage{
checkGRlist(myGRList)
}
\arguments{
\item{myGRList}{a list of GRanges objects}
}
\value{
TRUE if the input is a list of GRanges objects
}
\description{
Check that the given object is a list of GRanges objects
}
\details{
This function checks that the given object is a list of GRanges objects.
}
\examples{
require(GenomicRanges)
GR1=GRanges(seqnames="1", ranges=IRanges(start=1, end=1000))
GR2=GRanges(seqnames="1", ranges=IRanges(start=10, end=2000))
checkGRlist(list(GR1, GR2))
}
\author{
tlesluyes
}
