% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGR.R
\name{checkGRlist}
\alias{checkGRlist}
\title{checkGRlist}
\usage{
checkGRlist(myGRList, checkOverlaps = TRUE)
}
\arguments{
\item{myGRList}{a list of GRanges objects}

\item{checkOverlaps}{whether to check for overlapping ranges in the GRanges objects (default: TRUE)}
}
\value{
TRUE if the input is a list of valid GRanges objects
}
\description{
Check that the given object is a list of valid GRanges objects
}
\details{
This function checks that the given object is a list of valid GRanges objects.
}
\examples{
GR1 <- GenomicRanges::GRanges(seqnames="1", ranges=IRanges::IRanges(start=1, end=1000))
GR2 <- GenomicRanges::GRanges(seqnames="1", ranges=IRanges::IRanges(start=10, end=2000))
checkGRlist(list(GR1, GR2))
}
\author{
tlesluyes
}
