% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_load.R
\name{fun_load}
\alias{fun_load}
\title{Load Package}
\usage{
fun_load(..., char, install = TRUE, type = "binary", character.only = FALSE)
}
\arguments{
\item{...}{一个包名向量或字符串}

\item{char}{包名,字符串}

\item{install}{尝试安装库中找不到的包}

\item{type}{以二进制的形式安装, 或者 "source"}

\item{character.only}{如果为TRUE，则p_load将只接受单个字符串，默认接受向量字符串}
}
\value{
返回值为空, 不可见
}
\description{
加载包,如果该包没有安装,则安装以后再加载
}
\seealso{
\link[pacman:p_load]{pacman::p_load}
}
