% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA_paths.R
\name{get_all_paths}
\alias{get_all_paths}
\title{get_all_paths}
\usage{
get_all_paths(start, end, WGD, max_path_size = 5, simplify = TRUE)
}
\arguments{
\item{start}{a vector of length 2 (representing a copy-number state; e.g. c(1, 1) represents a 1+1 state), defining where to start}

\item{end}{a vector of length 2 (representing a copy-number state; e.g. c(1, 1) represents a 1+1 state), defining where to end}

\item{WGD}{a boolean defining if WGD events are allowed}

\item{max_path_size}{an integer defining the maximum path size}

\item{simplify}{a boolean defining if consecutive and opposite alterations (e.g. +1/+0 and then -1/-0) are allowed}
}
\value{
A vector of all possible paths given as characters (separator=";")
}
\description{
Get all possibles paths between two copy-number states
}
\details{
This function returns all possible paths between two copy-number states. The expected input is allele-specific (with two values), but it can be used for total copy-number by setting c(ntot, 0). Possible events include: +1/+0 (gain of the major allele), -1/-0 (loss of the major allele), +0/+1 (gain of the minor allele), -0/-1 (loss of the minor allele) and WGD.
}
\examples{
# Diploid baseline (1+1) turns into 2+1
print(get_all_paths(start=c(1, 1), end=c(2, 1), WGD=TRUE))
# Chromosome X in males (1+0) is gained (5 copies)
print(get_all_paths(start=c(1, 0), end=c(5, 0), WGD=TRUE))
}
\author{
tlesluyes
}
