% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{get_arms}
\alias{get_arms}
\title{get_arms}
\usage{
get_arms(assembly, withCentromeres = TRUE)
}
\arguments{
\item{assembly}{an assembly (e.g. hg38) or a data.frame with expected cytoband information (chr, start, end, name, gieStain)}

\item{withCentromeres}{whether to include centromeric regions (default: TRUE)}
}
\value{
A data.frame with genomic regions, can be converted to GRanges on the fly
}
\description{
Get chromosome arms information
}
\details{
This function gets chromosome arms information for a given assembly.
}
\examples{
get_arms("hg38")
}
\author{
tlesluyes
}
