% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoband_CHRsize.R
\name{get_centromeres}
\alias{get_centromeres}
\title{get_centromeres}
\usage{
get_centromeres(assembly)
}
\arguments{
\item{assembly}{an assembly (e.g. hg38) or a data.frame with expected cytoband information (chr, start, end, gieStain)}
}
\value{
A data.frame with genomic regions, can be converted to GRanges on the fly
}
\description{
Get centromeres information
}
\details{
This function gets centromeres information for a given assembly.
}
\examples{
get_centromeres("hg38")
}
\author{
tlesluyes
}
