% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA_paths.R
\name{get_shortest_path}
\alias{get_shortest_path}
\title{get_shortest_path}
\usage{
get_shortest_path(paths, wanted_WGD = NA, count_WGD = FALSE)
}
\arguments{
\item{paths}{all possible paths to consider}

\item{wanted_WGD}{a numeric value defining the number of WGD events wanted (can be NA to allow for any possibility, including no event at all; default: NA)}

\item{count_WGD}{a boolean defining if the number of WGD events should be counted (default: FALSE)}
}
\value{
A numeric value representing the minimal number of events, its name represents the full path
}
\description{
Get the shortest path among several
}
\details{
This function returns the shortest possible path. It should be used after running the get_all_paths function or can be used as long as the input format is correct.
}
\examples{
# Diploid baseline (1+1) turns into 2+1
print(get_shortest_path(get_all_paths(start=c(1, 1), end=c(2, 1), WGD=TRUE)))
# Chromosome X in males (1+0) is gained (5 copies)
print(get_shortest_path(get_all_paths(start=c(1, 0), end=c(5, 0), WGD=TRUE)))
}
\author{
tlesluyes
}
