% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeGRanges.R
\name{harmonizeGRanges}
\alias{harmonizeGRanges}
\title{harmonizeGRanges}
\usage{
harmonizeGRanges(myGRList, cores = 1)
}
\arguments{
\item{myGRList}{a list of GRanges objects, each object should correspond to one CNA profile}

\item{cores}{a numeric, the number of cores to use (default: 1)}
}
\value{
A list of harmonized GRanges objects
}
\description{
Harmonize GRanges objects
}
\details{
This function harmonizes GRanges objects by keeping only regions covered by all samples.
}
\examples{
require("GenomicRanges", "doParallel")
GR1=GRanges(seqnames="1", ranges=IRanges(start=1, end=1000), nMajor=1, nMinor=1)
GR2=GRanges(seqnames="1", ranges=IRanges(start=10, end=2000), nMajor=2, nMinor=1)
harmonizeGRanges(list(GR1, GR2))
stopImplicitCluster()
}
\author{
tlesluyes
}
