% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun-package.R
\docType{package}
\name{myfun-package}
\alias{myfun}
\alias{myfun-package}
\title{myfun: Small functions I think are easy to use}
\description{
Collect some small functions, mainly because they always forget the name of other people's packages, and only remember this function name.
}
\author{
\strong{Maintainer}: Zou shucai \email{zscmoyjian@163.com}

}
\keyword{internal}
