% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrenceGRanges.R
\name{occurrenceGRanges}
\alias{occurrenceGRanges}
\title{occurrenceGRanges}
\usage{
occurrenceGRanges(myGRList, myMetadata)
}
\arguments{
\item{myGRList}{a list of GRanges objects, each object should correspond to one CNA profile}

\item{myMetadata}{a vector of metadata to consider}
}
\value{
A GRanges object with nSamples as the total number of samples and metadata columns with the occurrence of events
}
\description{
Get the occurrence of events
}
\details{
This function gets the occurrence of events in a list of GRanges objects. All objects must have the same metadata columns and metadata must be TRUE/FALSE.
}
\examples{
GR1=GenomicRanges::GRanges(seqnames="1",
                           ranges=IRanges::IRanges(start=1, end=1000),
                           Gain=TRUE, Loss=FALSE)
GR2=GenomicRanges::GRanges(seqnames="1",
                           ranges=IRanges::IRanges(start=10, end=2000),
                           Gain=FALSE, Loss=TRUE)
occurrenceGRanges(list(GR1, GR2), c("Gain", "Loss"))
}
\author{
tlesluyes
}
