% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackages.R
\name{RpackageDependencies}
\alias{RpackageDependencies}
\title{RpackageDependencies}
\usage{
RpackageDependencies(
  customFolder = NULL,
  customDependencyTypes = NULL,
  customColours = NULL,
  simplifyNetwork = TRUE,
  saveFile = NULL
)
}
\arguments{
\item{customFolder}{a vector of folder names (default: NULL; .libPaths() is used)}

\item{customDependencyTypes}{a vector of dependency types, possible values are: "Depends", "Imports", "LinkingTo", "Suggests" and "Enhances" (default: c("Depends", "Imports", "LinkingTo"))}

\item{customColours}{a named vector of colours. Names must correspond to the dependency types and values must be valid colours (default: NULL; an internal colour scheme is used)}

\item{simplifyNetwork}{a boolean defining if the network should be simplified, i.e. the R base packages are removed (default: TRUE)}

\item{saveFile}{a string defining the name of the HTML file where the network should be saved (default: NULL; no file is saved)}
}
\value{
A list with nodes (a data.frame of R packages), links (a data.frame of package dependencies) and plot (a network plot using networkD3)
}
\description{
Show the package dependencies
}
\details{
Given a folder of R packages, this function reads the DESCRIPTION files of the installed packages and shows their dependencies.
}
\examples{
myDep=RpackageDependencies()
print(head(myDep$nodes))
print(head(myDep$links))
}
\author{
tlesluyes
}
