% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDF.R
\name{splitDF}
\alias{splitDF}
\title{splitDF}
\usage{
splitDF(DF, chunks)
}
\arguments{
\item{DF}{a data.frame to split}

\item{chunks}{a number of chunks to obtain}
}
\value{
A list of data.frames
}
\description{
Split a data.frame
}
\details{
This function splits a data.frame into a list of data.frames.
}
\examples{
DF=data.frame(a=1:26, b=letters)
splitDF(DF, 3)
}
\author{
tlesluyes
}
